requirements.txt
# Requirements for Healthcare Skill Intelligence System
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Data Processing
pandas==2.3.3
numpy==2.4.1

# HTTP Client
requests==2.31.0
httpx==0.25.1

# Utilities
python-multipart==0.0.6
